+++ packages/asc-ui/src/components/List/OrderedList.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import OrderedListStyle from './OrderedListStyle'
+
+const OrderedList: React.FC<React.HTMLAttributes<HTMLUListElement>> = ({
+  children,
+  ...otherProps
+}) => {
+  return <OrderedListStyle {...otherProps}>{children}</OrderedListStyle>
+}
+
+export default OrderedList